requirements.txt
streamlit
mistralai
faiss-cpu
numpy
python-dotenv
requests